requirements.txt
streamlit
Pillow
python-dotenv
requests
pillow-heif  # For HEIC files from iPhone
webptools  # For WebP support
